+++ examples/basic-usage/commands/b/c.js
@@ -2,5 +2,6 @@ module.exports = {
   command: function (args, flags, context) {
     console.log('c', args, flags, context)
   },
-  options: []
+  args: [],
+  flags: []
 }