+++ examples/basic-usage/index.js
@@ -1,17 +1,8 @@
 #! /usr/bin/env node
 
-var path = require('path')
-var directoryCommand = require('../../index')
+const path = require('path')
+const directoryCommand = require('../../index')
 
-var directory = path.join(__dirname, 'commands')
+const directory = path.join(__dirname, 'commands')
 
-var defaultCommand = {
-  command: function () {},
-  options: []
-}
-
-var options = {
-  defaultCommand
-}
-
-directoryCommand(directory, process.argv.slice(2), options)
+directoryCommand(directory, process.argv.slice(2))