+++ tests/fixtures/simple/a.js
@@ -1,6 +1,18 @@
+function command (args, flags, context) {
+  context.t.ok(flags.hi === 'huh')
+}
+
+const args = []
+
+const flags = [
+  {
+    name: 'hi',
+    type: 'string'
+  }
+]
+
 module.exports = {
-  command: function (args, flags, context) {
-    context.t.ok(flags.hi === 'huh')
-  },
-  options: []
+  command,
+  args,
+  flags
 }