+++ tests/fixtures/simple/b/c.js
@@ -2,5 +2,11 @@ module.exports = {
   command: function (args, flags, context) {
     context.t.ok(flags.hi === 'huh')
   },
-  options: []
+  args: [],
+  flags: [
+    {
+      name: 'hi',
+      type: 'string'
+    }
+  ]
 }